ADD: Pause to sys.exit calls
This should fix issues where users clicking the exe will have it flash
shortly then close, without being able to read the output.

Also fix some improper formatting and issue with unicode characters.  If
a game has a special character in it, it will now throw an exception and
just be listed as "app " + appID