Made start.py multi-platform.
Replaced the Windows-only command os.system('pause') with universal
command raw_input().
Extended function idleOpen() for Linux and Mac. The function now
creates a global object for the started idling process, so that
the function idleClose() can use the universal method terminate()
to stop the idling process.
Added a command to switch to the folder of the script/binary to fix
a bug in OS X.