UPDATE: delete browser control cookie data on load
The form that contains the authentication browser control will now
always delete the session and cookie data that is saved from it's
previous use or from Internet Explorer.  This will make the process of
using Idle Master for multiple accounts easier as a user will no longer
be automatically signed in using their previous authentication data.