Adding new cookies for the session not being overthrown by web browser session.
The idle master has a bug that, after it had logged in, its session
could be reset by Steam server if you go to Steam store (or community)
from web browser. You can login in web browser or already be logged in.
After you open up any page in browser, the idle master will lose its
session and ask to sign in again. But the browser doesn't lose its
session. It is because of "steamMachineAuth" cookie it has that is given
on first connection. With this cookie the idle master will not lose its
session.