ADD: Beginning of localization
This is the beginning of localization for Idle Master.  This commit
includes three sample English and Chinese strings, but has support for
more localizations as required.  Loading string elements in the
Form_Load() for each form will be the normal way of accomplishing this
rather than having a separate resource file for each form.