ADD: Increased resiliency when not connected
Previously, if the computer running Idle Master was disconnected from
Steam (either their local connection was down or Steam's servers were
offline) then the program would log the user off.  With this, Idle
Master will retry every 10 seconds for a total of 3 minutes to
re-establish a connection with Steam before doing so.  This should
resolve a few issues with users being "kicked off" when Steam's servers
randomly return status 503 Unavailable.

This also alters the behavior of StartIdle so that all instances of
steam-idle processes are killed before starting new processes.